migration changes

+++ database/migrations/2019_10_16_153946_create_status_types_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('status_types', function (Blueprint $table) {
             $table->smallInteger('id')->autoIncrement();
             $table->string('status_name', 20)->unique();
-            $table->smallInteger('status_code')->unique();
+            $table->tinyInteger('status_code')->unique();
             $table->timestamps();
         });
     }
